refactor: utils에서 gameResult 분리
